1st refactor attempt (#1)
* Update gitignore, group files into folders

* Refactor currency view out of content view

* Refactor vocab view out of content view

* Group main views into folders
